🔨 Refactor(#118): 폴더구조 변경
