+++ src/features/quiz/ui/Result.tsx
@@ -6,8 +6,8 @@ import useUserStore from '@store/useUserStore';
 import type { PartStatus } from '@features/learn/types';
 import type { Quiz } from '@features/quiz/types';
 import { isLoggedIn } from '@features/user/service/authUtils';
-import { isCompleted } from '@features/quiz/service/quizUtils';
-import { useHpUpdate } from '@/features/quiz/service/hooks';
+import { isCompleted } from '@/features/quiz/utils';
+import { useHpUpdate } from '@/features/quiz/hooks';
 
 interface ResultProps {
   quizId: Quiz['id'];
